ecs: cut over spawning, players+petals+bots, and the client
Three stages from a multi-agent workflow. Spawning and players were
adversarially reviewed; the client was NOT — all three client agents died on
an auth error after the implementation had landed, so that stage carries no
review at all.

SPAWNING (both reviewers: ship)
Mobs are now BORN as ECS entities via server/enemyRegistry.ts rather than
built as Enemy literals and imported every tick. The per-tick mob import pass
in ecsSync is gone; what remains is the removal half, because ~14 legacy sites
still splice enemies[]. Orphan adoptions (a shell with no entity) are counted
and logged rather than silently handled — an orphan means some path bypassed
the registry, which would present as "a few mobs are statues".

PLAYERS + PETALS + BOTS (both reviewers: fix-first — see below)
Player movement integration moved to the ECS, deliberately narrow: the ECS
decides where a flower's velocity carries it and nothing else. Petals, pickups,
teleporters, respawn and collision stay legacy. configureCutover now THROWS if
a disabled system name matches nothing on any scheduler — the same silent
no-op shape as the projectile bug, made fatal at boot instead.

CLIENT (unreviewed)
game.ts's two entity Maps are gone; graphics/** and net/handlers/** read the
ECS world. Verified mechanically: zero remaining references to game.enemies or
game.players. The client layer gained real render state — IsPlayerRender,
NeedsSnap, MobRender, DpsLabel — and a LegacyPlayer component holding the ~45
account/UI fields that do not belong in columns. Position, facing, eased eye
and the petal anchor were DELETED from the Player interface so no renderer can
read a stale copy. Interpolation now derives its ease rate from the user's
interpolationAmount setting rather than a constant, and snaps rather than
glides past TELEPORT_SNAP_DISTANCE (the maze is 200000px away).

OUTSTANDING — NOT FIXED HERE
- PVP knockback is discarded: applyPvpDamage writes victim.x += knockbackX
  from inside the ATTACKER's updatePlayerState, i.e. after the victim's move.
- A yggdrasil revive sets otherPlayer.isDead = false from inside another
  player's updatePlayerState, after the move.
  Both are the write-outside-the-sync-window failure this rewrite keeps
  reproducing.
- updatePlayerPreMovement hoists three passes (sponge damage, passive healing,
  raindrop aura) that do not feed movement, changing their order.
- The spawning removal path and the whole admission path are untested.
- The entire client stage is unreviewed and untested beyond typecheck.

Gates: both tsconfigs clean, npm run test:ecs passes, harness stable at
1.969ms mean / 7.197ms worst, zero non-finite coordinates. One harness run
showed a 107ms worst tick that did not reproduce; the harness asserts the MEAN
only, so a real spike would not fail it — that is a gap in the gate, not a
clean bill of health.

dist/ deliberately not rebuilt.

+++ src/ecs/components/player.ts
@@ -44,11 +44,23 @@ export const PlayerInput = defineComponent('PlayerInput', {
     /** string[] of held keys. */
     keys: 'obj',
     useMouse: 'bool',
-    /** Normalised mouse direction, -1..1. */
-    mouseDirectionX: 'f32',
-    mouseDirectionY: 'f32',
+    /**
+     * Normalised mouse direction, -1..1.
+     *
+     * f64, not f32, even though the magnitudes are tiny. These three fields are
+     * multiplied by MAX_SPEED and the speed factor to produce the target
+     * velocity that the SHARED `stepPlayerMovement` integrates, and the legacy
+     * path this replaces did that arithmetic in doubles. Storing them narrower
+     * would make the ECS result differ from the legacy one in the low bits, and
+     * that difference is not observable in the game (petal positions are
+     * quantised to 3-6px before they are hashed for the broadcast) but it IS
+     * observable in the cutover check, which asserts exact equality against the
+     * legacy formula. Keeping the check exact is worth 12 bytes per player.
+     */
+    mouseDirectionX: 'f64',
+    mouseDirectionY: 'f64',
     /** Client-computed speed multiplier, 0.15..1.0. */
-    mouseSpeedMultiplier: 'f32',
+    mouseSpeedMultiplier: 'f64',
     petalExtension: 'f32',
 });
 
@@ -137,10 +149,17 @@ export const Cosmetics = defineComponent('Cosmetics', {
  * client so client-side prediction moves at exactly the server's speed.
  */
 export const PlayerModifiers = defineComponent('PlayerModifiers', {
-    speedBoost: 'f32',
-    speedFactor: 'f32',
+    /**
+     * The three fields movement consumes are f64 for the same reason
+     * PlayerInput's direction fields are: they feed the shared physics step, and
+     * the legacy code they replace computed them in doubles. `sizeMultiplier`
+     * additionally decides the substep count inside `stepPlayerMovement`, where
+     * a rounding difference can change how many substeps a wall contact gets.
+     */
+    speedBoost: 'f64',
+    speedFactor: 'f64',
     /** Multiplier on the flower's radius and hitbox. */
-    sizeMultiplier: 'f32',
+    sizeMultiplier: 'f64',
     /** Additive pixels on item pickup radius. */
     magnetism: 'f32',
     /** Additive pixels on the range at which mobs aggro this player. */
